requirements.txt
# requirements.txt

# Core ML/AI Libraries
llama-index==0.10.57
chromadb==0.4.24
smolagents==0.1.0
mistralai==0.4.0
google-generativeai==0.7.2

# Data Processing
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.28
requests==2.31.0
beautifulsoup4==4.12.2

# Vector Embeddings
sentence-transformers==2.2.2
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-chroma==0.1.7
llama-index-llms-mistralai==0.1.4

# Visualization
matplotlib==3.7.2
plotly==5.17.0
seaborn==0.12.2

# Web Framework
streamlit==1.28.0
flask==2.3.3

# Text Processing
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Async & HTTP
aiohttp==3.8.6
asyncio==3.4.3

# Configuration
python-dotenv==1.0.0
pydantic==2.4.2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# News API
newsapi-python==0.2.7
praw==7.7.1  # Reddit API

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2